+++ core/designsystem/src/main/java/com/suwiki/core/designsystem/component/bottomsheet/SuwikiAlignBottomSheetContent.kt
@@ -1,4 +1,4 @@
-package com.suwiki.feature.lectureevaluation.viewerreporter.component
+package com.suwiki.core.designsystem.component.bottomsheet
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
@@ -11,26 +11,21 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.suwiki.core.designsystem.component.align.SuwikiAlignContainer
-import com.suwiki.core.designsystem.component.bottomsheet.SuwikiBottomSheet
 import com.suwiki.core.designsystem.theme.Gray95
 import com.suwiki.core.designsystem.theme.SuwikiTheme
-import com.suwiki.feature.lectureevaluation.viewerreporter.LectureEvaluationState
-import com.suwiki.feature.lectureevaluation.viewerreporter.R
-import kotlinx.collections.immutable.PersistentList
-import kotlinx.collections.immutable.persistentListOf
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun AlignBottomSheet(
+fun SuwikiAlignBottomSheet(
   showFilterSelectionBottomSheet: Boolean,
   hideAlignBottomSheet: () -> Unit,
   onClickAlignBottomSheetItem: (String) -> Unit,
+  itemList: List<String>,
+  bottomSheetTitle: String,
 ) {
-  var selectedItem by remember { mutableStateOf(-1) }
+  var selectedItem by remember { mutableStateOf(1) }
   SuwikiBottomSheet(
     sheetState = rememberModalBottomSheetState(
       skipPartiallyExpanded = true,
@@ -38,27 +33,22 @@ fun AlignBottomSheet(
     isSheetOpen = showFilterSelectionBottomSheet,
     onDismissRequest = hideAlignBottomSheet,
   ) {
-    AlignBottomSheetContent(
+    SuwikiAlignBottomSheetContent(
       selectedItem = selectedItem,
       onClickAlignBottomSheetItem = onClickAlignBottomSheetItem,
-      bottomSheetTitle = stringResource(R.string.word_sort),
+      bottomSheetTitle = bottomSheetTitle,
+      itemList = itemList,
     )
   }
 }
 
 @Composable
-private fun AlignBottomSheetContent(
+fun SuwikiAlignBottomSheetContent(
   selectedItem: Int,
   onClickAlignBottomSheetItem: (String) -> Unit = {},
   bottomSheetTitle: String,
+  itemList: List<String>,
 ) {
-  val filterDataList: PersistentList<String> = persistentListOf(
-    "최근 올라온 강의",
-    "꿀 강의",
-    "만족도 높은 강의",
-    "배울게 많은 강의",
-    "BEST 강의",
-  )
   Column(
     modifier = Modifier
       .padding(top = 36.dp, bottom = 45.dp),
@@ -69,7 +59,7 @@ private fun AlignBottomSheetContent(
       color = Gray95,
       modifier = Modifier.padding(horizontal = 24.dp, vertical = 15.dp),
     )
-    filterDataList.forEachIndexed { index, item ->
+    itemList.forEachIndexed { index, item ->
       val isChecked = selectedItem == index
       SuwikiAlignContainer(
         text = item,
@@ -79,13 +69,5 @@ private fun AlignBottomSheetContent(
         },
       )
     }
-  }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun FilterSelectionBottomSheetContentPreview() {
-  SuwikiTheme {
-    AlignBottomSheetContent(bottomSheetTitle = stringResource(R.string.word_sort), selectedItem = -1)
   }
 }
